Audio are now loaded only at application startup. ResourceName encapsulated in Reources
