+++ src/main/scala/model/nodes/util/RandomStoryNodeGenerator.scala
@@ -76,14 +76,6 @@ object RandomStoryNodeGenerator {
         StatEvent(StatModifier(StatName.Strength, i => i + 5))))
   }
 
-  private def setEnemy(): Option[Enemy] = {
-    val statValue: Int = 5
-    val maxPossibleHealth: Int = 100
-    val stats: Set[Stat] = Set(Stat(statValue, StatName.Speed), Stat(statValue, StatName.Defence))
-    rnd(EnemyProbability) match {
-      case 1 => Some(Enemy("jojo", rnd(maxPossibleHealth), stats))
-      case _ => None
-    }
-  }
+  
 
 }
