+++ src/main/scala/view/mainMenu/MainMenu.scala
@@ -2,11 +2,12 @@ package view.mainMenu
 
 import controller.ApplicationController
 import controller.ApplicationController.{isProgressAvailable, loadStoryNewGame, loadStoryWithProgress}
-import controller.util.ResourceName
+import controller.util.Resources.ResourceName
 import view.AbstractView
 import view.util.common.{ControlsPanel, Scrollable, VerticalButtons}
 import view.util.scalaQuestSwingComponents.{SqSwingButton, SqSwingLabel}
 import view.util.scalaQuestSwingComponents.dialog.SqYesNoSwingDialog
+
 import java.awt.BorderLayout
 import java.awt.event.ActionEvent
 import javax.swing.SwingConstants
