+++ src/main/scala/view/util/SoundPlayer.scala
@@ -1,22 +1,22 @@
 package view.util
 
-import controller.util.ResourceName.{interactionSoundEffectPath, navigationSoundEffectPath, resourceAsAudioInputStream}
+import controller.util.Resources.AudioClip.{interactionSoundClip, navigationSoundClip}
 
-import javax.sound.sampled.{AudioInputStream, AudioSystem, Clip}
+import javax.sound.sampled.Clip
 
 object SoundPlayer {
 
-  private def playAudio(audioInputStream: AudioInputStream): Unit = {
-    val clip : Clip = AudioSystem.getClip()
-    clip.open(audioInputStream)
+  def playClip(clip: Clip ): Unit = {
+    clip.setFramePosition(0)
     clip.start()
   }
 
   def playNavigationSound(): Unit = {
-    playAudio(resourceAsAudioInputStream(navigationSoundEffectPath()))
+    playClip(navigationSoundClip)
+
   }
 
   def playInteractionSound(): Unit = {
-    playAudio(resourceAsAudioInputStream(interactionSoundEffectPath()))
+    playClip(interactionSoundClip)
   }
 }
