+++ src/test/scala/controller/util/DirectoryInitializerTest.scala
@@ -1,9 +1,9 @@
 package controller.util
 
-import ResourceName.{gameDirectoryPath, storyDirectoryPath}
 import controller.util.DirectoryInitializer.StoryPopulationStrategy.TestStoryPopulation
 import controller.util.DirectoryInitializer.initializeGameFolderStructure
-import controller.util.ResourceName.MainDirectory.TempDirectory
+import controller.util.Resources.ResourceName.MainDirectory.TempDirectory
+import controller.util.Resources.ResourceName._
 import controller.util.serialization.FolderUtil.createFolderIfNotPresent
 import org.scalatest.DoNotDiscover
 import specs.FlatTestSpec
